Remove time.Sleep HACK
I replaced the time.Sleep HACK with a snippet I found in
https://github.com/vipul-sharma20/midi-macro. midi-macro is
permissively licensed so copying it is okay.

Since I am yet to read the chapters on goroutines and channels, I did
not understand how exactly this worked, but I asked him about it in
the issue tracker. I made a comment refering to the issie before each
instance of the snippet.

This is not perfect, since I don't understand how it works, but it is
much better than the time.Sleep HACK preceeding it.